Update jw_ptv.c
Working on calibration with Claudio I have encountered a problem to get detections in the Gimp edited, clean pics with very plain background and very bright dots. The older versions didn't have high pass in the calibration procedure and worked with these photos, but not this version. I suggest to comment out the highpass and always clean and prepare the images for the calibration.